Reverting the addition of 1709 as a valid OS flag value
After internal discussion, it was decided to keep the original OS
version as 2016 and not introduce 1709 to the users. This may confuse
the users as the final stemcell being produced by both versions would be
worded as 2016.

There is a lot of documentation and warnings that already inform the
user that a 2016 Windows stemcells use Windows 2016v1709

[#161622565]