Renamed package command line flag from 'version' to 'stemcell-version'
We found the original package command line flag not a true indication of what it
was used for. In addition, we want a flag to return the application's
version, and for that we wanted the 'version' flag. So, we renamed the
flag to something more accurate and free up 'version' for future
implementation to return the apps version.

Also backfilled some test to ensure the package command line flags store
the values in the appropriate member variable. This should help us
future proof any changes to command line flags (i.e. wanting to rename
them).

[#161989688]