Added version information to help text
We want the application to show its version information in the default
help text, and the help text shown by the `help` command.

The Concourse semver file is now stored in this repo, which is then go
generated into a go file that is used by the help text to show the
version.

[#161882286]

Signed-off-by: Gab Satchi <gsatchithanantham@pivotal.io>
Co-authored-by: Gab Satchi <gsatchithanantham@pivotal.io>