Fix integration tests to work on Windows
On Windows, Go cannot build a binary if the output file already exist.
The Go TempFile not only generates a random temoporary file name, it also
creates the file on the system. And since that was the filename used to
generate the output of the go build, go failed on the build.

In addition, on Windows, executable files *MUST* end with ".exe" or
".com" extention; without those, Windows will not consider the files
executable.

[#162520942]