Ensure there is enough free space before creating stemcell
Since creating a stemcell can be a long process, if there is not enough
free space on the system, stembuild will fail and the user will need to
start over. As such, we now check to ensure there is enough free space
before creating the stemcell, thus reducing the amount of time wasted
creating it.

Stembuild ensures there's at least twice the size of the VMDK before
creating the stemcell, otherwise it will fail early.

[#160354077](https://www.pivotaltracker.com/story/show/160354077)

Signed-off-by: Jackson Feeny <jfeeny@pivotal.io>
Co-authored-by: Jackson Feeny <jfeeny@pivotal.io>