Runs "quicker" Dism cleanup in ProvisionVM
During manual testing, we discovered that _maybe_ running Dism.exe
before we do the rest of the work in the `ProvisionVM` function might
cause the later, more thorough Dism.exe call to not fail with a variety
of failures.

Let's call Dism.exe before and after we do that work and see if it helps
at all.

[#180628580] Promote Windows 2019 based stemcells - 2021-12-14

Signed-off-by: Kenneth Lakin <klakin@vmware.com>