Run `Dism.exe` 3 times, not 8 times
We don't want to run `Dism.exe` more times than is necessary. Originally
we ran it only two times until 2021-12-15, after which it
consistently failed. We added six more `Dism.exe` runs throughout the
code, but we feel that's a lazy solution, and with this commit we're
attempting to minimize the number of times we invoke `Dism.exe`.

Note:

- Prior to 2021-12-15, the windows-2019-stemcell /
  create-stembuild-linux-stemcell / run-stembuild-construct ( pipeline /
  job / task ) took approximately 65 minutes to complete, after
  increasing Dism.exe to eight runs it took 110 minutes to complete
- Prior to 2021-12-15, the first run of Dism.exe took approximately 24
  minutes to run, after, 65 minutes.

[#180628580]