🐞 Ensure BOSH.SSH Install-SSHD func is available
The automation helper script that we use to setup the stemcell makes a
call to a custom method we've written called Install-SSHD in the
BOSH.SSH module. This conflicts with the recently introduced
`install-sshd.ps1` script that OpenSSH 9.1 installs.

Normally, calling Install-SSHD would cause Powershell to look through
its list of modules and load the appropriate one. However, since OpenSSH
9.1 now puts its `install-sshd.ps1` script in a default path, that
  script was being executed instead of causing the module lookup.

We believe this is only a problem for our tests, because that
Install-SSHD function is the way we install OpenSSH, and thus the
conflicting script does not exist when we call the function during the
stemcell build process.

Fixes error in
https://ci.bosh-ecosystem.cf-app.com/teams/main/pipelines/windows-2019-stemcell/jobs/stembuild-windows/builds/802:
```
A parameter cannot be found that matches parameter name 'SSHZipFile'.
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>