Update `stembuild` binary to accept "setup args" ...
... which can be passed to the powershell execution.

This is done in service of setting a flag so that the window stemcell
creation can fail if the windows update CA Cert can not be loaded.

The fail-on-CA-cert installation should not happend by default since
many customers expect to build windows stemcells without access to the
ouside world (including the windows update certificate providers).

Signed-off-by: aram price <pricear@vmware.com>