Add a flag to fail if Window Update Cert installation fails
This required modifying the `Setup.ps1` script, `function Setup()`, and
`function ProvisionVM()` so that the place where the exception is
normally supressed is allowe to fail.

This has been added because in testing we alway expect the Window Update
certificates to be installed, however the validation comes much later in
the testing cycle. This change allows us to see clearly _why_ expected
certificates are not present.

This change does modify the default behavior since users of `stembuild`
may expect the WU certificate installation to fail (gracefully) in cases
where the stemcell is being build without external network access.

Signed-off-by: aram price <pricear@vmware.com>