Run Dism.exe in a loop waiting for it to succeed
We saw failures in some environments where Dism.exe fails due to "locked resources"
which we believe is a different instance of Dism.exe already running. We are not exactly sure
what is triggering the background run of Dism.exe, but simply waiting for it to finish
seems like the cleanest solution.