Update construction of StemcellAutomation.zip
This updates the construction of StemcellAutomation.zip to require
user-provided files. This is closer to how the official binaries for
`stembuild` are built, and should be safer than the previous way assets
were being downloaded. A script has been added to build the zip file
when provided with the appropriate assets.