Explicitly wait for stembuild to exit
We are seeing frequent flakes in CI where the construct suite will
partially run, but then exit without an explicit error, causing the
suite to fail. We somewhat suspect it's related to our usage of
gexec.Start and the session it returns.

Now, we explicitly wait for the session to exit before continuing to the
next test.

We've also removed a spec that seemed to be checking that stembuild runs
for at least 40 seconds when the target VM exists. This does not appear
to provide value, and was one of the specs that it would start running
but then never exit.

Signed-off-by: Brian Upton <brian.upton@broadcom.com>